% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_orthonormal_matrix}
\alias{generate_orthonormal_matrix}
\title{Generate a Random Orthonormal Matrix}
\usage{
generate_orthonormal_matrix(p)
}
\arguments{
\item{p}{The dimensionality of the desired matrix.}
}
\value{
A \code{p}-by-\code{p} orthonormal matrix.
}
\description{
Creates an orthonormal matrix based on the QR decomposition of a square
matrix with standard Gaussian entries.
}
\keyword{internal}
