% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{generate_rfg_params}
\alias{generate_rfg_params}
\title{Generate Parameters for an RFG Function}
\usage{
generate_rfg_params(
  p,
  q = 1L,
  num.funs = 20L,
  a.min = -1,
  a.max = 1,
  lambda = 2,
  d.min = 0.1,
  d.max = 2
)
}
\arguments{
\item{p}{The dimensionality of the RFG function's domain.}

\item{q}{The output dimensionality.}

\item{num.funs}{The number of Gaussian basis functions to sum. Corresponds
to the parameter \eqn{L} in the original paper.}

\item{a.min, a.max}{The lower and upper bounds for the uniform distribution
from which the linear coefficients \code{a} are drawn.}

\item{lambda}{The rate parameter for the exponential distribution used to
determine the latent dimensionality of each basis function.}

\item{d.min, d.max}{The lower and upper bounds for the uniform distribution
from which the singular values of the covariance matrices \code{V} are drawn.}
}
\value{
An \code{list} of \verb{<rfg_params>}, each specifying the parameters for a
single scalar-valued RFG function.
}
\description{
Generate Parameters for an RFG Function
}
\details{
Generates a collection of complete parameter sets for a (potentially
vector-valued) RFG function.

We define a vector-valued RFG function as a collection of
independently-generated scalar-valued RFG functions. Specifically, if
\eqn{F^*(x) \in \mathbb R^q} is the \eqn{q}-dimensional output of a
vector-valued RFG function \eqn{F^*:\mathbb R^p\to\mathbb R^q}, then
\deqn{F^*(x) = (F_1^*(x),\ldots,F_q^*(x))}
where each \eqn{F_j^*(x) \in \mathbb R} is the scalar output of a
scalar-valued RFG function \eqn{F_j^*:\mathbb R^p\to\mathbb R}.
}
\references{
Friedman, J. H. (2001). Greedy function approximation: A gradient boosting machine.
The Annals of Statistics, 29(5), 1189-1232.
\doi{10.1214/aos/1013203451}
}
\seealso{
\code{\link[=rfg]{rfg()}}
}
