% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{generate_scalar_rfg_params}
\alias{generate_scalar_rfg_params}
\title{Generate Parameters for a Scalar-Valued RFG Function}
\usage{
generate_scalar_rfg_params(p, num.funs, a.min, a.max, lambda, d.min, d.max)
}
\arguments{
\item{p}{The dimensionality of the RFG function's domain.}

\item{num.funs}{The number of Gaussian basis functions to sum. Corresponds
to the parameter \eqn{L} in the original paper.}

\item{a.min, a.max}{The lower and upper bounds for the uniform distribution
from which the linear coefficients \code{a} are drawn.}

\item{lambda}{The rate parameter for the exponential distribution used to
determine the latent dimensionality of each basis function.}

\item{d.min, d.max}{The lower and upper bounds for the uniform distribution
from which the singular values of the covariance matrices \code{V} are drawn.}
}
\value{
An object of class \verb{<rfg_params>}.
}
\description{
Generates a complete set of parameters for a scalar-valued RFG function.
}
\references{
Friedman, J. H. (2001). Greedy function approximation: A gradient boosting machine.
The Annals of Statistics, 29(5), 1189-1232.
\doi{10.1214/aos/1013203451}
}
\seealso{
\code{\link[=rfg]{rfg()}}
}
\keyword{internal}
