% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Extract Parameters from an RFG Function}
\usage{
params(x, ...)
}
\arguments{
\item{x}{An object of class \verb{<rfg_function>}.}

\item{...}{Other arguments passed to \code{params()}. Currently ignored.}
}
\value{
A \code{list} of \verb{<rfg_params>} objects.
}
\description{
An S7 generic to extract the \code{list} of \verb{<rfg_params>} from a generated
\verb{<rfg_function>} object.
}
