% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfg-package.R
\docType{package}
\name{rfg-package}
\alias{rfg}
\alias{rfg-package}
\title{rfg: Random Function Generator}
\description{
Generates complex multivariate functions following the random function generator design used in Friedman (2001) 'Greedy function approximation: A gradient boosting machine'. Intended for use as part of data pipeline for synethic data experiments.
}
\author{
\strong{Maintainer}: David Fleischer \email{david.p.fleischer@gmail.com}

}
\keyword{internal}
