% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfg.R
\name{rfg}
\alias{rfg}
\title{Create a RFG Function}
\usage{
rfg(p, q = 1L, ..., seed = NULL)
}
\arguments{
\item{p}{The primary dispatch argument. Can be either:
\itemize{
\item A positive integer indicating the desired input dimensionality \code{p}.
\item A \code{list} of \verb{<rfg_params>}, each representing the parameters for
an individual scalar-valued RFG function. The dimensionality of the RFG
function's range is implied by the length of the \verb{<rfg_params>} list.
}}

\item{q}{When \code{p} is an integer value, \code{q} is the desired dimensionality
of the RFG function's range (output). Defaults to \code{q = 1L} and generates
a scalar-valued RFG function. Otherwise, when \code{p} is a \verb{<rfg_params>}
object, \code{q} is ignored.}

\item{...}{When \code{p} is an integer value, these are optional arguments
passed to \code{\link[=generate_rfg_params]{generate_rfg_params()}}. Otheriwse ignored.}

\item{seed}{When \code{p} is an integer value, an optional seed may be supplied
for reproducible parameter generation. See the internal function
\code{\link[=with_optional_seed]{with_optional_seed()}} for details.}
}
\value{
An object of class \verb{<rfg_function>}.
}
\description{
Creates a RFG function by either randomly generating the RFG parameters, or
by supplying a \code{list} of \verb{<rfg_params>} objects, each specifying the
component functions of a vector-valued RFG function.
}
\details{
This function is an S7 generic that can be called as \code{rfg(p, ...)} in the
following ways:
\itemize{
\item Automatically generating parameters: When \code{p} is a positive integer,
call \code{rfg(p, ...)} to generate a new RFG function whose domain (input)
has dimensionality \code{p}.
\item Pre-specified parameters: When \code{p} is a \code{list} of \verb{<rfg_params>}
objects, call \code{rfg(p)} to create the corresponding RFG function.
}
}
\references{
Friedman, J. H. (2001). Greedy function approximation: A gradient boosting machine.
The Annals of Statistics, 29(5), 1189-1232.
\doi{10.1214/aos/1013203451}
}
\seealso{
\code{\link[=generate_rfg_params]{generate_rfg_params()}}, \code{\link[=with_optional_seed]{with_optional_seed()}}
}
