% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{with_optional_seed}
\alias{with_optional_seed}
\title{Execute Code with an Optional Random Seed}
\usage{
with_optional_seed(code, seed)
}
\arguments{
\item{code}{The expression to evaluate.}

\item{seed}{An optional numeric seed.}
}
\description{
A simple wrapper around \code{\link[withr:with_seed]{withr::with_seed()}}, designed to avoid messages
caused by a missing or \code{NULL} seed.
}
\keyword{internal}
