% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_orthonormal_matrix}
\alias{generate_orthonormal_matrix}
\title{Generate a Random Orthonormal Matrix}
\usage{
generate_orthonormal_matrix(p)
}
\arguments{
\item{p}{The dimensionality of the desired matrix.}
}
\value{
A \code{p}-by-\code{p} orthonormal matrix.
}
\description{
Creates an orthonormal matrix based on the QR decomposition of a square
matrix with standard Gaussian entries.
}
\details{
TODO ... the reason we generate orthonormal matrices this way is because
the distribution of the matrix \eqn{Q} from the QR decomposition on the
IID Gaussian matrix \eqn{M = QR} will be uniform over the space of
orthonormal matrices \eqn{Q \in O(p)} ... this means that the resulting
matrix \eqn{Q} can be thought of as a random rotation/reflection matrix
that is unbiased over the space of possible rotation/reflection matrices
in the sense of having no preferred rotational direction ...
}
\keyword{internal}
