% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfg-classes.R
\name{rfg_basis_params}
\alias{rfg_basis_params}
\title{The \code{rfg_basis_params} Class}
\usage{
rfg_basis_params(
  a = integer(0),
  phi = integer(0),
  mu = integer(0),
  V = integer(0)
)
}
\description{
An S7 class to represent a parameter set of an individual basis function.
}
\keyword{internal}
