% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{rfg_function}
\alias{rfg_function}
\title{The \code{rfg_function} Class}
\usage{
rfg_function(..., params)
}
\description{
An S7 class to represent a callable RFG function returned by \code{\link[=rfg]{rfg()}}. It
inherits from \code{S7::class_function} and holds the list of \verb{<rfg_params>}
objects that specify an RFG function.
}
\keyword{internal}
