% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfg-classes.R
\name{rfg_params}
\alias{rfg_params}
\title{The \code{rfg_params} Class}
\usage{
rfg_params(p = integer(0), bases = list())
}
\description{
An S7 class to represent a complete set of parameters for specifying a
scalar-valued RFG function. An object of class \verb{<rfg_params>} holds the
dimensionality \code{p} of an RFG function's domain and a \code{list} of
\verb{<rfg_basis_params>}, a collection of parameters necessary to specify each
of the basis functions that defines a scalar-valued RFG function.
}
\seealso{
\code{\link[=rfg_basis_params]{rfg_basis_params()}}
}
\keyword{internal}
